+++ src/filter/date.js
@@ -1,5 +1,7 @@
+var phpDate = require('locutus/php/datetime/date');
+
 module.exports = function(value, params) {
-  var date = Twig.functions.date(value);
+  var date = phpDate(value);
   var format = params && params.length ? params[0] : 'F j, Y H:i';
-  return Twig.lib.date(format, date);
+  return phpDate(format, date);
 }