+++ src/filter/format.js
@@ -1,7 +1,9 @@
+var vsprintf = require('locutus/php/strings/vsprintf');
+
 module.exports = function(value, params) {
   if (value === undefined || value === null){
     return;
   }
 
-  return Twig.lib.vsprintf(value, params);
+  return vsprintf(value, params);
 }