+++ src/filter/striptags.js
@@ -1,7 +1,9 @@
+var strip_tags = require('locutus/php/strings/strip_tags');
+
 module.exports = function(value) {
   if (value === undefined || value === null){
     return;
   }
 
-  return Twig.lib.strip_tags(value);
+  return strip_tags(value);
 }